+++ __tests__/integration/harness.ts
@@ -1,3 +1,3 @@
-const endPoint = "http://localhost:3000";
+const endPoint = "http://localhost:3100";
 
 export default endPoint;