+++ test_harness/imports/client/main.js
@@ -1,0 +1,2 @@
+console.log("Client main");
+document.write('Client main');